attempt to fix issue with url on remote deployment with subpath
